+++ include/co/net/status_codes.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <system_error>
+#include <co/status_code.hpp>
 
 namespace co::net::impl
 {
